Fix initialize of project
Fix init database because flask-sqlalchemy was
updated. Now it's prohibited to call db.init_app(app)
more that one time.

Fix basic auth init
