+++ SpiderKeeper/app/spider/model.py
@@ -1,6 +1,15 @@
 import datetime
 from sqlalchemy import desc, text
-from SpiderKeeper.app import db, Base
+from SpiderKeeper.app import db
+
+
+class Base(db.Model):
+    __abstract__ = True
+
+    id = db.Column(db.Integer, primary_key=True)
+    date_created = db.Column(db.DateTime, default=db.func.current_timestamp())
+    date_modified = db.Column(db.DateTime, default=db.func.current_timestamp(),
+                              onupdate=db.func.current_timestamp())
 
 
 class Project(Base):
